requirements.txt
Flask
python-dotenv
SQLAlchemy
langchain
langchain-openai
langgraph
typing_extensions
email-validator
scikit-learn
numpy
openai